backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.7.14
click==8.2.1
distro==1.9.0
Flask==3.1.1
flask-cors==6.0.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
numpy==2.3.1
openai==1.97.1
pandas==2.3.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
Werkzeug==3.1.3
openai-whisper
torch --index-url https://download.pytorch.org/whl/cpu
speechbrain>=0.5.15
torchaudio>=2.0.0